vertex-ai/cc-fraud-model/vertex-model-training/requirements.txt
pandas
scikit-learn
gcsfs==0.6.2
google-cloud-storage==1.30.0
